feat!: generate can now extend using special "latest" value
Extending by specifying "latest" will check if only one file with
"filename" exists, and use that.

An option `-u` (unique) is also added which is what is now expanded to a
date and what the `-x` (extend) option compares against. `-o` (output)
is now rather the actual file name we save to.